cleaned up files and made a shared service for components to use
