--- src/app/app.component.html
@@ -1,9 +1,0 @@
-<!-- app.component.html -->
-<app-header></app-header> <!-- Header and navigation -->
-
-<router-outlet></router-outlet> <!-- This is where routed components will be displayed -->
-
-<app-footer></app-footer> <!-- Footer -->
-
-<app-error-comp></app-error-comp> <!-- Error handling and messages -->
-
