+++ src/app/header/header.component.html
@@ -13,7 +13,7 @@
         <input
           id="searchForm"
           type="text"
-          class="flex-grow px-4 py-2 border rounded-2xl focus:outline-none focus:ring-2 focus:ring-gray-500 focus:border-transparent"
+          class="flex-grow px-4 py-2 border rounded focus:outline-none focus:ring-2 focus:ring-gray-500 focus:border-transparent"
           placeholder="Search for a City"
           [formControl]="searchQuery"
           [matAutocomplete]="auto"
