--- src/app/search.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SearchService } from './search.service';
-
-describe('SearchService', () => {
-  let service: SearchService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SearchService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
