--- src/app/search.service.ts
@@ -1,22 +1,0 @@
-import { Injectable } from '@angular/core';
-import { BehaviorSubject, Observable } from 'rxjs';
-
-@Injectable({
-  providedIn: 'root'
-})
-
-export class SearchService {
-
-  private searchQuerySubject: BehaviorSubject<string> = new BehaviorSubject<string>('');
-  public searchQuery$: Observable<string> = this.searchQuerySubject.asObservable();
-
-
-  setSearchQuery(query: string) {
-    this.searchQuerySubject.next(query.toLocaleLowerCase());
-  }
-
-  getSearchQuery(): string {
-    return this.searchQuerySubject.value;
-  } 
-
-}
